libtt/drummachine: Fix signed division bug
TT_DRUMMACHINE_VOICE_LENGTH is used in division expressions on signed
numbers. It must therefore be signed to ensure these expressions don't
perform unsigned division (thus not propagating the sign bit).

Curiously this bug cannot be observed on 64-bit systems where the
expression converts to 64-bit unsigned division and a later downcast to
32-bit restores the sign bit!

+++ src/libtt/drummachine.c
@@ -25,7 +25,7 @@
 #define TT_DRUMMACHINE_DIVISIONS_PER_BEAT 1
 #define TT_DRUMMACHINE_PATTERN 2
 
-#define TT_DRUMMACHINE_VOICE_LENGTH lengthof(kick_Dry_b)
+#define TT_DRUMMACHINE_VOICE_LENGTH ((int) lengthof(kick_Dry_b))
 
 uint8_t basic[] = {
 	4,					// length of pattern
